fix(backwards-compatible): adjustments so kong-build-tools is backwards compatible and a CI check to verify as such (#560)
* fix(backwards-compatible): adjustments so we're backwards compatible with 2.8.0

* fix(tests): keep the tests just make them conditional

* chore(ci): make our build process less verbose [reverts 92e4d4b]

* fix(syntax): derpy syntax fix

+++ test/tests/01-package/run.sh
@@ -19,6 +19,9 @@ DOCKER_SYSTEM_ARCHITECTURE="${DOCKER_SYSTEM_ARCHITECTURE:-$(
   uname -m
 )}"
 
+# architecture in the kong package filename (amd64 or arm64) default: amd64
+KONG_ARCHITECTURE='amd64'
+
 case "_${DOCKER_SYSTEM_ARCHITECTURE}" in
   _aarch64|_arm64)
     BASE_DOCKER_PLATFORM='linux/arm64/v8'
@@ -42,7 +45,7 @@ docker run \
   "$IMAGE_BASE" \
   tail -f /dev/null
 if [[ "$PACKAGE_TYPE" == "rpm" ]]; then
-  cp $PACKAGE_LOCATION/*amd64.rpm kong.rpm
+  cp $PACKAGE_LOCATION/*${KONG_ARCHITECTURE}.rpm kong.rpm
   docker exec ${USE_TTY} user-validation-tests /bin/bash -c "yum install -y /src/kong.rpm procps"
   docker exec ${USE_TTY} user-validation-tests /bin/bash -c "kong version"
 # Tests disabled until CSRE-467 is resolved
@@ -51,7 +54,7 @@ if [[ "$PACKAGE_TYPE" == "rpm" ]]; then
 fi
 
 if [[ "$PACKAGE_TYPE" == "deb" ]]; then
-  cp $PACKAGE_LOCATION/*amd64.deb kong.deb
+  cp $PACKAGE_LOCATION/*${KONG_ARCHITECTURE}.deb kong.deb
   docker exec ${USE_TTY} user-validation-tests /bin/bash -c "apt-get update"
   docker exec ${USE_TTY} user-validation-tests /bin/bash -c "apt-get install -y perl-base zlib1g-dev"
   docker exec ${USE_TTY} user-validation-tests /bin/bash -c "dpkg -i /src/kong.deb || apt install --fix-broken -y"
@@ -156,15 +159,19 @@ docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "grep 
 # kong shipped files
 docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "ls -l /etc/kong/kong.conf.default"
 docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "ls -l /etc/kong/kong*.logrotate"
-docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "ls -l /usr/local/kong/include/kong/pluginsocket.proto"
-docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "ls -l /usr/local/kong/include/wrpc/wrpc.proto"
 docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "ls -l /usr/local/kong/include/google/protobuf/*.proto"
+
+if [ -e ${KONG_SOURCE_LOCATION}/kong/include/kong/pluginsocket.proto ]; then
+  docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "ls -l /usr/local/kong/include/kong/pluginsocket.proto"
+fi
+if [ -e ${KONG_SOURCE_LOCATION}/kong/include/wrpc/wrpc.proto ]; then
+  docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "ls -l /usr/local/kong/include/wrpc/wrpc.proto"
+fi
 
 if [[ "$EDITION" == "enterprise" ]]; then
   docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /usr/local/openresty/bin/resty -e 'require("ffi").load "passwdqc"'
   docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /usr/local/openresty/bin/resty -e 'require("ffi").load "jq"'
   #docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} openapi2kong 2>&1 | head -1 | grep 'missing required parameter:'
-  docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} ls -l /usr/local/kong/include/kong/pluginsocket.proto
   docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/bash -c "ls -l /usr/local/kong/include/google/protobuf/*.proto"
 fi
 
@@ -176,7 +183,7 @@ fi
 
 # TODO enable this test in other distros containing systemd
 if [[ "$RESTY_IMAGE_BASE" == "ubuntu" ]] && [ -z "${DARWIN:-}" ]; then
-  cp $PACKAGE_LOCATION/*amd64.deb kong.deb
+  cp $PACKAGE_LOCATION/*${KONG_ARCHITECTURE}.deb kong.deb
   docker run -d --rm --name systemd-ubuntu -e KONG_DATABASE=off --privileged -v /sys/fs/cgroup:/sys/fs/cgroup:ro -v $PWD:/src jrei/systemd-ubuntu:$RESTY_IMAGE_TAG
   docker exec ${USE_TTY} systemd-ubuntu /bin/bash -c "apt-get clean"
   docker exec ${USE_TTY} systemd-ubuntu /bin/bash -c "apt-get update"
